requirements.txt
Django>=4.2,<5.0
Pillow>=10.0
whitenoise>=6.6
psycopg2-binary>=2.9
python-dotenv>=1.0
cryptography>=41.0.0
